Barley: ephemeral bare metal provisioning
Build two images with Packer:
- Seed is the bare metal boot image that includes debootstrap minbase,
  kernel, ssh, and small set of diagnostic tools;
- Sower is the PXE boot server container with dnsmasq in DHCP proxy
  configuration to serve pxelinux over TFTP, and nginx to serve the Seed
  kernel and initramfs over HTTP.

Sower is designed to run on a network that already has a DHCP server. It
obtains its own IP configuration from DHCP and leaves it up to the
existing DHCP server to allocate IP addresses to PXE clients.

Seed starts with all its physical Ethernet interfaces bound to a Linux
bridge br0. Seed build expects to find id_ed25519.pub in build directory
and puts it in authorized_keys for root. The root account is
passwordless and authenticating with SSH is the only way to access a
Seed host.

Limitations:
- Seed doesn't report back on boot, you'll have to scan your network or
  check your DHCP server's logs to find its IP address, or configure
  your DHCP server to allocate a static IP address based on its MAC.
- Sower only implements BIOS compatible network boot for now, UEFI
  compatible network boot will be added later.