Provision authorized_keys to Seeds from Sower
Instead of baking authorized_keys into the Seed initramfs image, send it
to the Seed along with other certificates when Seeds register with
Sower. The SSH public key to be used as the admin key is specified when
a new field is created, and defaults to ~/.ssh/id_ed25519.pub.